# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/auth/register":{"post":{"operationId":"AuthController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/reviews":{"post":{"operationId":"ReviewController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["Review"]},"get":{"operationId":"ReviewController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Review"]}},"/reviews/{id}":{"get":{"operationId":"ReviewController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]},"patch":{"operationId":"ReviewController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReviewDto"}}}},"responses":{"200":{"description":""}},"tags":["Review"]},"delete":{"operationId":"ReviewController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]}},"/service-providers":{"post":{"operationId":"ServiceProvidersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceProviderDto"}}}},"responses":{"201":{"description":""}},"tags":["ServiceProviders"]},"get":{"operationId":"ServiceProvidersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ServiceProviders"]}},"/service-providers/by-type":{"get":{"operationId":"ServiceProvidersController_findByType","parameters":[{"name":"type","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ServiceProviders"]}},"/service-providers/{id}":{"get":{"operationId":"ServiceProvidersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ServiceProviders"]},"patch":{"operationId":"ServiceProvidersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceProviderDto"}}}},"responses":{"200":{"description":""}},"tags":["ServiceProviders"]},"delete":{"operationId":"ServiceProvidersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ServiceProviders"]}},"/services":{"post":{"operationId":"ServiceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Service"]},"get":{"operationId":"ServiceController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Service"]}},"/services/by-service-provider":{"get":{"operationId":"ServiceController_findByServiceProviderId","parameters":[{"name":"serviceProviderId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Service"]}},"/services/{id}":{"get":{"operationId":"ServiceController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Service"]},"patch":{"operationId":"ServiceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["Service"]},"delete":{"operationId":"ServiceController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Service"]}}},"info":{"title":"API Documentation","description":"Generated API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateAuthDto":{"type":"object","properties":{}},"CreateReviewDto":{"type":"object","properties":{}},"UpdateReviewDto":{"type":"object","properties":{}},"CreateServiceProviderDto":{"type":"object","properties":{}},"UpdateServiceProviderDto":{"type":"object","properties":{}},"CreateServiceDto":{"type":"object","properties":{}},"UpdateServiceDto":{"type":"object","properties":{}}}}}